Fixed all warnings

+++ src/circle.cpp
@@ -7,8 +7,8 @@ void Circle::generateCircle(glm::vec3 normal, glm::vec3 center, float radius, un
     // 3. y = (sqrt(D))/(2(a^2+b^2))
     // 4. x = -by/a
     float x{},y{},z{};
-    float D = 4*pow(normal.x,2)*pow(radius,2)*(pow(normal.x,2)+pow(normal.y,2));
-    y = pow(D,0.5)/(2*(pow(normal.x,2)+pow(normal.y,2)));
+    float D =(float) (4*pow(normal.x,2)*pow(radius,2)*(pow(normal.x,2)+pow(normal.y,2)));
+    y = (float)(pow(D,0.5)/(2*(pow(normal.x,2)+pow(normal.y,2))));
     x = -normal.y*y/normal.x;
     
     // 5. Get the coordinates of the point by adding x1,y1and z1 to x,y and z.
